Moving navigation config into AppViewModel
The reasoning for this is that this is the navigation
model for this microservice.
